Putting constrains on weight size.
Summary:
Making weight size cosntrained on N and K. N % 16 == 0 and K % 4 == 0.
Without this shape bound propagation is breaking things and the only
solution is to breakup gemm + dequant op in two relay ops. For now
putting this constrain to move forward. May remove it in the future.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags: